refactor: Move payroll to its own package
