+++ soap24/client.go
@@ -6,13 +6,13 @@ import (
 	"fmt"
 
 	"github.com/hooklift/gowsdl/soap"
+	"github.com/valuetechdev/24sevenoffice-go/payroll24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/account24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/attachment24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/auth24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/client24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/company24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/invoice24"
-	"github.com/valuetechdev/24sevenoffice-go/soap24/payroll24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/person24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/product24"
 	"github.com/valuetechdev/24sevenoffice-go/soap24/project24"
